Renamed msvc installer, and added SP1 version, now that I applied SP1 on dev machine. Changed windows installer to use MSVC10SP1 redistributable. Avoided several MSVC compiler warnings in NeuronAnnotator code.

+++ v3d_main/neuron_annotator/data_model/CustomClipPlanes.cpp
@@ -88,7 +88,7 @@ void CustomClipPlanes::simpleUndoPlane(int planeIndex)
     // qDebug() << "removing plane" << planeIndex + 1;
     int topPlaneIndex = nextClipPlaneIndex - 1;
     if (topPlaneIndex < 0)
-        topPlaneIndex = size() - 1;
+        topPlaneIndex = (int)size() - 1;
     if (planeIndex == topPlaneIndex)
         nextClipPlaneIndex = topPlaneIndex;
     (*this)[planeIndex].disable();
